Split base factory helpers
