+++ tensordict/_base/__init__.py
@@ -1,0 +1,34 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+#
+# This source code is licensed under the MIT license found in the
+# LICENSE file in the root directory of this source tree.
+
+"""Internal implementation modules for :mod:`tensordict.base`."""
+
+from tensordict._base.factories import (
+    from_any,
+    from_csv,
+    from_dict,
+    from_h5,
+    from_json,
+    from_list,
+    from_namedtuple,
+    from_pandas,
+    from_parquet,
+    from_struct_array,
+    from_tuple,
+)
+
+__all__ = [
+    "from_any",
+    "from_csv",
+    "from_dict",
+    "from_h5",
+    "from_json",
+    "from_list",
+    "from_namedtuple",
+    "from_pandas",
+    "from_parquet",
+    "from_struct_array",
+    "from_tuple",
+]
